Move extension methods to its own syntax object
